feat(api): reorganizar estrutura de arquivos da API e atualizar rotas para adoption e hero dog
