+++ api/hero-dog/get.ts
@@ -1,8 +1,8 @@
-import { kv } from "./_lib/kv";
+import { kv } from "../_lib/kv";
 import type { IncomingMessage, ServerResponse } from "http";
-import { validateRequest } from "./_lib/validation";
-import { sendError, sendSuccess } from "./_lib/response";
-import { HTTP_STATUS } from "./_lib/constants";
+import { validateRequest } from "../_lib/validation";
+import { sendError, sendSuccess } from "../_lib/response";
+import { HTTP_STATUS } from "../_lib/constants";
 
 export default async function handler(
   _req: IncomingMessage,
