+++ vercel.json
@@ -7,7 +7,7 @@
   ],
   "crons": [
     {
-      "path": "/api/update-hero-dog",
+      "path": "/api/hero-dog/update",
       "schedule": "0 0 * * *"
     }
   ]
